update after folder reorganization * change various paths * adds editors back into app.js Part of #59
