+++ lib/app.js
@@ -3,6 +3,7 @@ const Feed = require('rss'); // to create feed objects and
 const sanitize = require('sanitize-filename');
 const feedmerger = require('./feedmerger.js').feedmerger;
 const pagewriter = require('./pagewriter.js').pagewriter;
+const editors = require('./editors.js');
 
 // helper functions
 
@@ -44,7 +45,7 @@ const createSidebar = function (categories) {
     sidebar += '<a href="{{ site.baseurl }}/' + url + '.html">' + category.replace(/&/g, '&amp;') + '</a>\n';
   }
   sidebar += '</nav>';
-  fs.writeFile('./mathblogging.org/_includes/sidebar-secondary.html', sidebar, function (err) {
+  fs.writeFile('../mathblogging.org/_includes/sidebar-secondary.html', sidebar, function (err) {
     'use strict';
     if (err) {
       console.log('error: couldn\'t write Sidebar');
@@ -65,7 +66,7 @@ const createBlogIndex = function (blogs) {
     blogIndex += '  <li>\n    <a href="' + blog.url + '" rel="nofollow"> ' + blog.url + '</a>\n  </li>\n';
   }
   blogIndex += '</ul>\n';
-  fs.writeFile('./mathblogging.org/blogindex.html', blogIndex, {
+  fs.writeFile('../mathblogging.org/blogindex.html', blogIndex, {
     mode: 0o664
   }, function (err) {
     if (err) {
@@ -88,10 +89,11 @@ const mergeFeedsOfCategory = function (category, feedFilenames) {
   });
   feedmerger(feedFilenames, categoryFeed, pagewriter.bind(null, category));
 }
+
 // The action
 
 // load the main json file
-const FeedsJson = JSON.parse(fs.readFileSync('./feeds.json', 'utf8'));
+const FeedsJson = JSON.parse(fs.readFileSync('../data/feeds.json', 'utf8'));
 
 // extract categories
 const categories = filterCategories(FeedsJson);
@@ -105,3 +107,6 @@ createBlogIndex(FeedsJson.blogs);
 for (let category of categories) {
   mergeFeedsOfCategory(category, feedFilenamesForCategory(FeedsJson, category))
 }
+
+// generate Editors' Picks
+editors();
