+++ lib/editors.js
@@ -7,7 +7,7 @@ var Feed = require('rss');
 var async = require('async');
 
 // set up Twitter app secrets
-var fileName = './secret-config.json';
+var fileName = '../secret-config.json';
 var config = {};
 
 try {
@@ -113,7 +113,7 @@ module.exports = function() {
     var xml = editorFeed.xml({
       indent: true
     });
-    fs.writeFile('./mathblogging.org/editors-picks.xml', xml, function(err) {
+    fs.writeFile('../mathblogging.org/editors-picks.xml', xml, function(err) {
       if (err) {
         console.log('error: couldn\'t write Editors\' Picks Feed');
         return console.log(err);
